Add sample dApp
The dApp is able to connect to both Browser Wallet and WalletConnect. It includes a network selector for switching between testnet and mainnet, and a contract selector for getting information about a contract. Apart from UI rendering, the vast majority of code is located in libraries.

The next PR will include the ability to invoke/update the selected smart contract via the wallet connection.

A "network inconsistency detector" is included to surface the problems with getting the network from the browser wallet.

Includes a lot of addition and fixes in the libraries.