react-components: Remove 'connectActive' method (#15)
Remove the responsibility of initiating connections from `WithWalletConnector`. This simplifies this component and takes away any confusion as to why there's more than one way to establish a connection (and what exactly each method does and doesn't).

Instead of using the `connectActive` method, one should just call `connect` directly on the connector which will then notify `WithWalletConnector` via its role as the connector's delegate. The application then calls `setActiveConnection` to set it as active if desired.

This means that `activeConnectorError` no longer holds errors for connections (only the connector's error as the name also implies) and that the `isConnecting` indicator no longer is maintained: It's now the application's job to track this state.