react-components: Extract 'activeConnection' into hook (#20)
Remove `activeConnection` and its friends (`activeConnectedAccount`, `activeConnectionGenesisHash`, and `setActiveConnection`) and move their combined state into a new hook `useConnection`.

The field `connectedAccounts` now maps connections to the empty string if they don't have an associated account. This means that the key set of this field matches the set of live connections exactly, which is important for downstream state to be able to detect disconnections and clean themselves up appropriately.