wallet-connector: WalletConnect: Implement binary message signing
If the message is "binary" it's encoded into hex and the schema provided in an additional field (encoded as base64). If the wallet doesn't read the schema, it's just going to present it non-decoded.

Unrelated change: Fixed helper functions in sample dapps for extracting the message of an error - WalletConnect has a habit of throwing errors with empty message, meaning that the error is erroneously type cast to a string.