wallet-connectors: WalletConnect: Use "lower" camelCase format for transaction type (#42)
This apparently is the standard style used for these strings. This is
achieved by using the new function `getTransactionKindString` in
`@concordium/common-sdk@9.0.0`. We use a yarn resolution to override the
version specified by `@concordium/browser-wallet-api-helpers`
(https://github.com/Concordium/concordium-browser-wallet/pull/324
suggests reducing this dependency to a "peer dependency").

This doesn't change anything in the Android wallet as it doesn't read
the field but just assumes that it's an `update` transaction. The iOS
implementation reads the field and supports both formats for backwards
compatibility with projects using old versions of
`@concordium/wallet-connectors`. The Android implementation must do the
same when it starts supporting other transaction types.

Resolves [CBW-1145](https://concordium.atlassian.net/browse/CBW-1145).

[CBW-1145]:
https://concordium.atlassian.net/browse/CBW-1145?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ