Fix executable on Windows
`ps1` and `cmd` files are automatically generated from this script on Windows, but it interprets `-S` as the main program to run instead of `node`, which means that the scripts don't work.
To get around it, I've changed the script so that it calls node without
`--no-warnings` (and thus `-S`), but then it spawns a child process of
`node --no-warnings` which executes the actual cli file.